feat: feature flagged all of 3d
