+++ Cargo.toml
@@ -11,14 +11,11 @@ keywords = ["bevy"]
 exclude = ["assets/*"]
 
 [features]
-# ## Enables 3d render target implementation
-# 3d-support = ["bevy/bevy_pbr"]
+## Enables 3d render target implementation
+3d = ["bevy/bevy_pbr", "bevy/bevy_mesh_picking_backend"]
 
-# ## By default, enabled 3d
-# default = ["3d"]
-
-## For internal use only
-internal-debugging = ["bevy/track_change_detection"]
+## By default, enabled 3d
+default = ["3d"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -33,8 +30,6 @@ bevy = { version = "0.15", default-features = false, features = [
     "bevy_ui",
     "bevy_winit",
     "bevy_window",
-    "bevy_pbr",
-    "bevy_mesh_picking_backend",
     "png",
     "x11",
     "webgl2",
@@ -65,3 +60,7 @@ web-sys = { version = "0.3.70", features = [
 [dev-dependencies]
 insta = "1.29.0"
 bevy = "0.15"
+
+[[example]]
+name = "basic_3d"
+required-features = ["3d"]
