+++ src/impls/scan.rs
@@ -9,6 +9,7 @@ use crate::prelude::*;
 pub enum SourceType {
     Ui,
     Sprite,
+    #[cfg(feature = "3d")]
     ThreeD,
 }
 
@@ -17,6 +18,7 @@ pub enum SourceType {
 pub struct RenderTypeScan {
     is_sprite: Has<TextEdit2d>,
     is_ui: Has<TextEdit>,
+    #[cfg(feature = "3d")]
     is_3d: Has<TextEdit3d>,
 }
 
@@ -28,15 +30,28 @@ impl RenderTypeScanItem<'_> {
         }
     }
 
+    fn is_3d(&self) -> bool {
+        #[cfg(feature = "3d")]
+        let ret = self.is_3d;
+
+        #[cfg(not(feature = "3d"))]
+        let ret = false;
+
+        ret
+    }
+
     pub(in crate::impls) fn scan(&self) -> Result<SourceType> {
-        let flags = [self.is_ui, self.is_sprite, self.is_3d];
+        let flags = [self.is_ui, self.is_sprite, self.is_3d()];
         let count_true = flags.iter().filter(|x| **x).count();
         match count_true {
             0 => Err(RenderTargetError::NoTargetsAvailable),
             1 => match flags {
                 [true, false, false] => Ok(SourceType::Ui),
                 [false, true, false] => Ok(SourceType::Sprite),
+                #[cfg(feature = "3d")]
                 [false, false, true] => Ok(SourceType::ThreeD),
+                #[cfg(not(feature = "3d"))]
+                [false, false, false] => unreachable!(),
                 _ => unreachable!(),
             },
             _ => Err(RenderTargetError::MoreThanOneTargetAvailable),
