+++ src/impls/threed.rs
@@ -16,3 +16,39 @@ pub(in crate::impls) fn sync_mesh_and_size(
         mesh_component.0 = mesh_server.add(mesh);
     }
 }
+
+impl TextEdit3d {
+    pub fn new(rendering_size: Vec2) -> Self {
+        Self {
+            world_size: rendering_size,
+            auto_manage_mesh: true,
+        }
+    }
+}
+
+pub(super) fn default_3d_material(
+    mut world: bevy::ecs::world::DeferredWorld,
+    target: Entity,
+    _: bevy::ecs::component::ComponentId,
+) {
+    let current_handle = world
+        .get::<MeshMaterial3d<StandardMaterial>>(target)
+        .unwrap()
+        .0
+        .clone();
+    if current_handle == Handle::default() {
+        debug!("It appears no customization of a `TextEdit3d` material has been done, overwriting with a default");
+        let default_material = StandardMaterial {
+            base_color: Color::WHITE,
+            unlit: true,
+            ..default()
+        };
+        let default_handle = world
+            .resource_mut::<Assets<StandardMaterial>>()
+            .add(default_material);
+        world
+            .get_mut::<MeshMaterial3d<StandardMaterial>>(target)
+            .unwrap()
+            .0 = default_handle;
+    }
+}
